Organise ndef and ntag stuff
