+++ cmd/nfc/ndef_test.go
